backend/app/api/deps.py
from typing import Generator, Optional

from fastapi import Depends, HTTPException, status
from fastapi.security import OAuth2PasswordBearer
from jose import jwt, JWTError
from pydantic import ValidationError

from app.core.config import settings
from app.core.security import verify_password
from app.core.logging import logger

# This will be used when we implement user authentication
oauth2_scheme = OAuth2PasswordBearer(
    tokenUrl=f"{settings.API_V1_STR}/auth/login"
)

# This is a placeholder for future database session handling
def get_db() -> Generator:
    try:
        # Here we would initialize a database connection
        db = None
        yield db
    finally:
        # Here we would close the database connection
        pass

# This is a placeholder for future user authentication
async def get_current_user(token: str = Depends(oauth2_scheme)):
    try:
        # Here we would decode the JWT token and get the user
        # For now, we'll just return a placeholder
        return {"username": "test_user"}
    except (JWTError, ValidationError):
        raise HTTPException(
            status_code=status.HTTP_403_FORBIDDEN,
            detail="Could not validate credentials",
        ) 