backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.8
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
bcrypt==4.0.1
httpx==0.24.1
email-validator==2.0.0 